Exception handlers and Sequencer Savefiles
Alright so now, whereas the program would have thrown an error when
attempting to send a command without a selected port, it.... well,
throws another error, but it's the intentional kind where I yell at you
instead of the program.

More important than that is that sequences can now be saved! The
sequencing wasn't really practical before because your sequence would be
erased once the program was closed, but they can now be saved to files.
The sequencing feature is definitely awesome to have, and is quite
frankly the only thing that separates this program from the other
controller someone wrote in Processing.